Add Unit Test Example
The UserServiceTest.deleteUser_goldenPath() tests the delete user method in the service. If you review the test code, you'll see what is being tested.

The arrange section demonstrates me making the test pretent like the the userRepository.findById returns an empty optional.

The act section is invoking the method I am wanting to test.

The assert makes assertions and verifications that the code behaved the way I expected it to.

For example, because I forced the userRepository.findById to return an empty optional, I would expect that the code never interacts with the user repository afterwards, which is proven on lines 77 and 78 of the test.

As an explanation for the file format, this is using JUNIT5 and mockito together. The @Mock tells the test to create an empty shell of the dependency of the class that I am testing. This allows me to imitate behavior from it by using a when() statement. The @InjeckMocks annotation is just a magical annotation that instantiates the service for me, injecting the empty shell of a dependency automatically before the test begins, being invoked by openMocks