Named Template Use Example
This is an example of how to the named template.

It's called a namedTemplate because it allows you to pass namedParameters to it, preventing SQL injection, which is a huge security risk.

The BeanPropertyRowMapper is a generic class that can infer the properties of the class passed in as an argument and try to deserialize the DB resultSet into those objects. If the properties don't line up directly with the DB colum names, it is request ot create a custom class that extends or implements beanPropertyRowMapper.